+++ test/Integration/ApiGateway/GetStringAsyncShould.cs
@@ -1,0 +1,449 @@
+﻿using System;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using AwsSignatureVersion4.Integration.ApiGateway.Authentication;
+using AwsSignatureVersion4.Private;
+using Shouldly;
+using Xunit;
+
+namespace AwsSignatureVersion4.Integration.ApiGateway
+{
+    public class GetStringAsyncShould : ApiGatewayIntegrationBase
+    {
+        public GetStringAsyncShould(IntegrationTestContext context)
+            : base(context)
+        {
+        }
+
+        #region GetStringAsync(string, string, string, <credentials>)
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestStringAndMutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestStringAndImmutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveImmutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        #endregion
+
+        #region GetStringAsync(Uri, string, string, <credentials>)
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestUriAndMutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl.ToUri(),
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestUriAndImmutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl.ToUri(),
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveImmutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        #endregion
+
+        #region GetStringAsync(string, HttpCompletionOption, string, string, <credentials>)
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestStringAndHttpCompletionOptionAndMutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var completionOption = HttpCompletionOption.ResponseContentRead;
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl,
+                completionOption,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestStringAndHttpCompletionOptionAndImmutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var completionOption = HttpCompletionOption.ResponseContentRead;
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl,
+                completionOption,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveImmutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        #endregion
+
+        #region GetStringAsync(Uri, HttpCompletionOption, string, string, <credentials>)
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestUriAndHttpCompletionOptionAndMutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var completionOption = HttpCompletionOption.ResponseContentRead;
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl.ToUri(),
+                completionOption,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestUriAndHttpCompletionOptionAndImmutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var completionOption = HttpCompletionOption.ResponseContentRead;
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl.ToUri(),
+                completionOption,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveImmutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        #endregion
+
+        #region GetStringAsync(string, CancellationToken, string, string, <credentials>)
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestStringAndCancellationTokenAndMutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var ct = new CancellationToken();
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl,
+                ct,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestStringAndCancellationTokenAndImmutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var ct = new CancellationToken();
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl,
+                ct,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveImmutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public void AbortGivenCanceled(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var ct = new CancellationToken(true);
+
+            // Act
+            var task = HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl,
+                ct,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            task.Status.ShouldBe(TaskStatus.Canceled);
+        }
+
+        #endregion
+
+        #region GetStringAsync(Uri, CancellationToken, string, string, <credentials>)
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestUriAndCancellationTokenAndMutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var ct = new CancellationToken();
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl.ToUri(),
+                ct,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestUriAndCancellationTokenAndImmutableCredentials(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var ct = new CancellationToken();
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl.ToUri(),
+                ct,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveImmutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        #endregion
+
+        #region GetStringAsync(string, HttpCompletionOption, CancellationToken, string, string, <credentials>)
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestStringAndHttpCompletionOptionAndCancellationTokenAndMutableCredentals(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var completionOption = HttpCompletionOption.ResponseContentRead;
+            var ct = new CancellationToken();
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl,
+                completionOption,
+                ct,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestStringAndHttpCompletionOptionAndCancellationTokenAndImmutableCredentals(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var completionOption = HttpCompletionOption.ResponseContentRead;
+            var ct = new CancellationToken();
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl,
+                completionOption,
+                ct,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveImmutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        #endregion
+
+        #region GetStringAsync(Uri, HttpCompletionOption, CancellationToken, string, string, <credentials>)
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestUriAndHttpCompletionOptionAndCancellationTokenAndMutableCredentals(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var completionOption = HttpCompletionOption.ResponseContentRead;
+            var ct = new CancellationToken();
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl.ToUri(),
+                completionOption,
+                ct,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenRequestUriAndHttpCompletionOptionAndCancellationTokenAndImmutableCredentals(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var completionOption = HttpCompletionOption.ResponseContentRead;
+            var ct = new CancellationToken();
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                Context.ApiGatewayUrl.ToUri(),
+                completionOption,
+                ct,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveImmutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        #endregion
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenQuery(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var uriBuilder = new UriBuilder(Context.ApiGatewayUrl)
+            {
+                Query = "Param1=value1"
+            };
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                uriBuilder.Uri,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenOrderedQuery(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var uriBuilder = new UriBuilder(Context.ApiGatewayUrl)
+            {
+                Query = "Param1=Value1&Param1=value2"
+            };
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                uriBuilder.Uri,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+
+        [Theory]
+        [InlineData(IamAuthenticationType.User)]
+        [InlineData(IamAuthenticationType.Role)]
+        public async Task SucceedGivenUnorderedQuery(IamAuthenticationType iamAuthenticationType)
+        {
+            // Arrange
+            var uriBuilder = new UriBuilder(Context.ApiGatewayUrl)
+            {
+                Query = "Param1=value2&Param1=Value1"
+            };
+
+            // Act
+            var stringContent =  await HttpClient.GetStringAsync(
+                uriBuilder.Uri,
+                Context.RegionName,
+                Context.ServiceName,
+                ResolveMutableCredentials(iamAuthenticationType));
+
+            // Assert
+            stringContent.ShouldBe("Not sure, please fill in");
+        }
+    }
+}