+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ## [2.1.1] - 2021-09-07
 
+### :zap: Added
+
+- [#526](https://github.com/FantasticFiasco/aws-signature-version-4/issues/526) Add support for `HttpClient.GetStringAsync` (contribution by [@mungojam](https://github.com/mungojam))
+
 ### :syringe: Fixed
 
 - [#478](https://github.com/FantasticFiasco/aws-signature-version-4/issues/478) Add support for `HttpClient.Send` on .NET 5 (contribution by [@Timovzl](https://github.com/Timovzl))